# docs/openapi.yaml
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  title: Gen3Workflow
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /ga4gh/tes/v1/service-info:
    get:
      operationId: service_info
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Service Info Ga4Gh Tes V1 Service Info Get
                type: object
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Service Info
      tags:
      - GA4GH TES
  /ga4gh/tes/v1/tasks:
    get:
      operationId: list_tasks
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response List Tasks Ga4Gh Tes V1 Tasks Get
                type: object
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: List Tasks
      tags:
      - GA4GH TES
    post:
      operationId: create_task
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Create Task Ga4Gh Tes V1 Tasks Post
                type: object
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Create Task
      tags:
      - GA4GH TES
  /ga4gh/tes/v1/tasks/{task_id}:
    get:
      operationId: get_task
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          title: Task Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Get Task Ga4Gh Tes V1 Tasks  Task Id  Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get Task
      tags:
      - GA4GH TES
  /ga4gh/tes/v1/tasks/{task_id}:cancel:
    post:
      operationId: cancel_task
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          title: Task Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Cancel Task Ga4Gh Tes V1 Tasks  Task Id  Cancel Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Cancel Task
      tags:
      - GA4GH TES
...